Create README.md
This is where I keep snippets of code that I found useful (or at least thought it might come in handy at some point). It might appear amateur, noddy, or whatever, but when I had an idea off the back of something I learned, I wanted to write it somewhere in case I need it (like a scratch pad).